+++ openfisca_us/variables/irs/income/agi.py
@@ -1,0 +1,15 @@
+from openfisca_us.model_api import *
+
+
+
+
+class posagi(Variable):
+    value_type = float
+    entity = TaxUnit
+    label = "Positive AGI"
+    unit = USD
+    documentation = "Negative AGI values capped at zero"
+    definition_period = YEAR
+
+    def formula(tax_unit, period, parameters):
+        return max_(tax_unit("c00100", period), 0)
